refactor: add the separate performer module
