+++ applications/tari_base_node/src/commands/mod.rs
@@ -2,5 +2,6 @@ mod args;
 pub mod cli;
 pub mod command_handler;
 pub mod parser;
+pub mod performer;
 pub mod reader;
 mod status_line;
